requirements.txt
fastapi
uvicorn
langgraph
python-dotenv
twilio
jinja2
sqlite-utils
google-generativeai
